resume_ai/requirements.txt
streamlit
pandas
pdfplumber
python-docx
spacy